Time: 7 ms (5.44%), Space: 8.7 MB (6.79%) - LeetHub

+++ 0007-reverse-integer/0007-reverse-integer.c
@@ -14,5 +14,5 @@ int reverse(int x) {
         rev_num = rev_num * 10 + digit;
     }
     
-    return rev_num;
+    return (int)rev_num;
 }
